requirements.txt
streamlit==1.32.0
python-docx==1.1.0
pymupdf==1.23.25
google-generativeai==0.8.4
pandas==2.2.0
requests==2.31.0
pdfplumber==0.10.3
python-dotenv==1.0.1
